+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -9,7 +9,6 @@ import java.util.HashSet;
 import java.util.Set;
 
 @Data
-@Builder
 public class User {
     @NotNull
     @PositiveOrZero
